requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
matplotlib>=3.7
autogluon.tabular>=1.1.0
openml>=0.14.2
